add Tree Generator to test
